move service tests into services
